+++ src/Libraries/SmartRestaurant.Application/Sections/Commands/SectionsCommandsHandler.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using FluentValidation.Results;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using SmartRestaurant.Application.Common.Exceptions;
 using SmartRestaurant.Application.Common.Interfaces;
 using SmartRestaurant.Domain.Entities;
@@ -12,7 +13,7 @@ namespace SmartRestaurant.Application.Sections.Commands
     public class SectionsCommandsHandler :
         IRequestHandler<CreateSectionCommand, ValidationResult>,
         IRequestHandler<UpdateSectionCommand, ValidationResult>,
-        IRequestHandler<DeleteSectionCommand>
+        IRequestHandler<DeleteSectionCommand, ValidationResult>
     {
         private readonly IApplicationDbContext _context;
         private readonly IMapper _mapper;
@@ -34,11 +35,14 @@ namespace SmartRestaurant.Application.Sections.Commands
             return default;
         }
 
-        public async Task<Unit> Handle(DeleteSectionCommand request, CancellationToken cancellationToken)
+        public async Task<ValidationResult> Handle(DeleteSectionCommand request, CancellationToken cancellationToken)
         {
-            var section = await _context.Sections.FindAsync(request.SectionId).ConfigureAwait(false);
+            var validator = new DeleteSectionCommandValidator();
+            var result = await validator.ValidateAsync(request, cancellationToken).ConfigureAwait(false);
+            if (!result.IsValid) return result;
+            var section = await _context.Sections.FindAsync(request.CmdId).ConfigureAwait(false);
             if (section == null)
-                throw new NotFoundException(nameof(Section), request.SectionId);
+                throw new NotFoundException(nameof(Section), request.CmdId);
             _context.Sections.Remove(section);
             await _context.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
             return default;
@@ -49,11 +53,13 @@ namespace SmartRestaurant.Application.Sections.Commands
             var validator = new UpdateSectionCommandValidator();
             var result = await validator.ValidateAsync(request, cancellationToken).ConfigureAwait(false);
             if (!result.IsValid) return result;
-            var section = await _context.Sections.FindAsync(request.CmdId).ConfigureAwait(false);
+            var section = await _context.Sections.AsNoTracking()
+                .FirstOrDefaultAsync(s => s.SectionId == request.CmdId, cancellationToken: cancellationToken)
+                .ConfigureAwait(false);
             if (section == null)
                 throw new NotFoundException(nameof(Section), request.CmdId);
-            section.Name = request.Name;
-            section.MenuId = request.MenuId;
+            var entity = _mapper.Map<Section>(request);
+            _context.Sections.Update(entity);
             await _context.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
             return default;
         }
